requirements.txt
# --------------------
# Core LangChain (1.x)
# --------------------
langchain==1.2.0
langchain-community==0.4.1
langchain-core==1.2.4
langchain-text-splitters==1.1.0

# --------------------
# LLM Providers
# --------------------
langchain-anthropic==1.3.0
langchain-openai==1.1.6
langchain-huggingface==1.2.0

# --------------------
# Vector Database
# --------------------
langchain-chroma==1.1.0
chromadb==1.3.7

# --------------------
# Embeddings
# --------------------
sentence-transformers==5.2.0
transformers==4.57.3

# --------------------
# Data Processing
# --------------------
pandas==2.3.3
numpy>=2.1.0,<2.3.0
openpyxl==3.1.5

# --------------------
# Core Compatibility FIXES
# --------------------
pydantic>=2.11.2,<3.0.0
protobuf>=6.32.1,<7.0.0
typing-extensions>=4.12.0

# --------------------
# OpenTelemetry (Version-Locked)
# --------------------
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0

# --------------------
# Utilities
# --------------------
python-dotenv==1.2.1

# --------------------
# API & Serving
# --------------------
fastapi==0.126.0
uvicorn==0.38.0

# --------------------
# Testing
# --------------------
pytest==9.0.2

# --------------------
# Optional: Jupyter & Visualization
# --------------------
jupyter==1.1.1
matplotlib==3.9.3
seaborn==0.13.2

